refactor: 상품 그룹 컴포넌트로 분리
- page내부 컴포넌트의 추상화 레벨을 비슷하게 가져가기 위해 컴포넌트로 분리 후 내부로 로직 이동
